// 1xbts-web/src/lib/proto/an/v1/service.ts
export interface AnEventRecord {
  receivedMs: number;
  session?: HrpdSessionEvent | undefined;
  traffic?: HrpdTrafficEvent1 | undefined;
  access?: HrpdAccessEvent | undefined;
}
// ...
export const AnEventRecord: MessageFns<AnEventRecord> = {
  encode(message: AnEventRecord, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.receivedMs !== 0) {
      writer.uint32(8).uint64(message.receivedMs);
    }
    if (message.session !== undefined) {
      HrpdSessionEvent.encode(message.session, writer.uint32(18).fork()).join();
    }
    if (message.traffic !== undefined) {
      HrpdTrafficEvent1.encode(message.traffic, writer.uint32(26).fork()).join();
    }
    if (message.access !== undefined) {
      HrpdAccessEvent.encode(message.access, writer.uint32(34).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): AnEventRecord {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseAnEventRecord();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.receivedMs = longToNumber(reader.uint64());
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.session = HrpdSessionEvent.decode(reader, reader.uint32());
          continue;
        }
        case 3: {
          if (tag !== 26) {
            break;
          }

          message.traffic = HrpdTrafficEvent1.decode(reader, reader.uint32());
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.access = HrpdAccessEvent.decode(reader, reader.uint32());
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  fromJSON(object: any): AnEventRecord {
    return {
      receivedMs: isSet(object.receivedMs)
        ? globalThis.Number(object.receivedMs)
        : isSet(object.received_ms)
        ? globalThis.Number(object.received_ms)
        : 0,
      session: isSet(object.session) ? HrpdSessionEvent.fromJSON(object.session) : undefined,
      traffic: isSet(object.traffic) ? HrpdTrafficEvent1.fromJSON(object.traffic) : undefined,
      access: isSet(object.access) ? HrpdAccessEvent.fromJSON(object.access) : undefined,
    };
  },

  toJSON(message: AnEventRecord): unknown {
    const obj: any = {};
    if (message.receivedMs !== 0) {
      obj.receivedMs = Math.round(message.receivedMs);
    }
    if (message.session !== undefined) {
      obj.session = HrpdSessionEvent.toJSON(message.session);
    }
    if (message.traffic !== undefined) {
      obj.traffic = HrpdTrafficEvent1.toJSON(message.traffic);
    }
    if (message.access !== undefined) {
      obj.access = HrpdAccessEvent.toJSON(message.access);
    }
    return obj;
  },

  create(base?: DeepPartial<AnEventRecord>): AnEventRecord {
    return AnEventRecord.fromPartial(base ?? {});
  },
  fromPartial(object: DeepPartial<AnEventRecord>): AnEventRecord {
    const message = createBaseAnEventRecord();
    message.receivedMs = object.receivedMs ?? 0;
    message.session = (object.session !== undefined && object.session !== null)
      ? HrpdSessionEvent.fromPartial(object.session)
      : undefined;
    message.traffic = (object.traffic !== undefined && object.traffic !== null)
      ? HrpdTrafficEvent1.fromPartial(object.traffic)
      : undefined;
    message.access = (object.access !== undefined && object.access !== null)
      ? HrpdAccessEvent.fromPartial(object.access)
      : undefined;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;
// ...
export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  fromJSON(object: any): T;
  toJSON(message: T): unknown;
  create(base?: DeepPartial<T>): T;
  fromPartial(object: DeepPartial<T>): T;
}